feat: split Explorer into Events + State Views as top-level pages
- Navbar split into App section (Catalog, Cart) and Demo section
  (Admin, Events, State Views, Info) with visual separator
- Demo links styled subtler (grey) to distinguish from app
- /events page: live event stream with filters
- /state page: live state view projections
- Removed old /explorer route
