+++ 04-boundless-typescript/frontend/src/components/Navbar.css
@@ -53,3 +53,26 @@
   background: #e8f5e9;
   color: #27ae60;
 }
+
+.nav-separator {
+  width: 1px;
+  height: 28px;
+  background: #ddd;
+  margin: 0 8px;
+  align-self: center;
+}
+
+.nav-item.demo {
+  font-size: 13px;
+  color: #999;
+}
+
+.nav-item.demo:hover {
+  background: #f5f5f5;
+  color: #666;
+}
+
+.nav-item.demo.active {
+  background: #f0f0f0;
+  color: #555;
+}
