+++ 04-boundless-typescript/frontend/src/components/Navbar.tsx
@@ -31,13 +31,19 @@ export default function Navbar() {
           <NavLink to="/cart" className={({ isActive }) => isActive ? 'nav-item active' : 'nav-item'}>
             Cart{itemCount > 0 ? ` (${itemCount})` : ''}
           </NavLink>
-          <NavLink to="/admin" className={({ isActive }) => isActive ? 'nav-item active' : 'nav-item'}>
+
+          <span className="nav-separator" />
+
+          <NavLink to="/admin" className={({ isActive }) => isActive ? 'nav-item demo active' : 'nav-item demo'}>
             Admin
           </NavLink>
-          <NavLink to="/explorer" className={({ isActive }) => isActive ? 'nav-item active' : 'nav-item'}>
-            Explorer
+          <NavLink to="/events" className={({ isActive }) => isActive ? 'nav-item demo active' : 'nav-item demo'}>
+            Events
           </NavLink>
-          <NavLink to="/info" className={({ isActive }) => isActive ? 'nav-item active' : 'nav-item'}>
+          <NavLink to="/state" className={({ isActive }) => isActive ? 'nav-item demo active' : 'nav-item demo'}>
+            State Views
+          </NavLink>
+          <NavLink to="/info" className={({ isActive }) => isActive ? 'nav-item demo active' : 'nav-item demo'}>
             Info
           </NavLink>
         </div>
